move social login into its own component
