+++ src/containers/Login/Login.js
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux';
 import LoginPage from 'components/Login/LoginPage';
-import { loginUser, loginWithGoogle, loginWithFacebook } from 'actions/auth';
+import { loginUser } from 'actions/auth';
 import PropTypes from 'prop-types';
 
 const mapStateToProps = state => ({
@@ -10,9 +10,7 @@ const mapStateToProps = state => ({
 });
 
 const mapDispatchToProps = dispatch => ({
-  login: (email, password) => dispatch(loginUser(email, password)),
-  loginWithGoogle: () => dispatch(loginWithGoogle()),
-  loginWithFacebook: () => dispatch(loginWithFacebook())
+  login: (email, password) => dispatch(loginUser(email, password))
 });
 
 LoginPage.propTypes = {
