+++ src/containers/Login/SocialLogin.js
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux';
-import RegisterPage from 'components/Login/RegisterPage';
-import { signUp } from 'actions/auth';
+import SocialLogin from 'components/Login/SocialLogin';
+import { loginWithGoogle, loginWithFacebook } from 'actions/auth';
 import PropTypes from 'prop-types';
 
 const mapStateToProps = state => ({
@@ -10,13 +10,14 @@ const mapStateToProps = state => ({
 });
 
 const mapDispatchToProps = dispatch => ({
-  signUp: (email, password, displayName) => dispatch(signUp(email, password, displayName))
+  loginWithGoogle: () => dispatch(loginWithGoogle()),
+  loginWithFacebook: () => dispatch(loginWithFacebook())
 });
 
-RegisterPage.propTypes = {
+SocialLogin.propTypes = {
   isLoggingIn: PropTypes.bool.isRequired,
   signUpError: PropTypes.bool,
   isAuthenticated: PropTypes.bool.isRequired
 };
 
-export default connect(mapStateToProps, mapDispatchToProps)(RegisterPage);
+export default connect(mapStateToProps, mapDispatchToProps)(SocialLogin);
